Move logic to helpers for reusability, fix errors
