+++ lib/helpers.js
@@ -9,6 +9,7 @@ var request = require('request');
 var installTimeoutSeconds = 30; //seconds to get a install response from device
 var uploadEndpoint = 'v2/app/resources/uploadurl';
 var appDetectFile = 'config.yaml';
+var prompt = require('prompt');
 
 
 var configFile = require('os').homedir() + '/.matrix/store.json';
@@ -875,6 +876,125 @@ function zipAppFolder(source, destination, cb) {
   zip.pipe(destinationZip); // send zip to the file
 }
 
+
+var profile = {
+  show: function (profile, callback) {
+    var err;
+    Matrix.loader.stop();
+    if (_.isUndefined(profile)) { //If no data
+      console.log(t('matrix.account.no_profile'));
+      callback();
+    } else {  
+      console.log('Name: '.yellow, profile.name);
+      console.log('URL: '.yellow, profile.url);
+      console.log('GitHub: '.yellow, profile.githubUrl);
+      console.log('Usage statistics: '.yellow, profile.trackOk ? 'Enabled'.green : 'Disabled'.yellow);
+      callback();
+    }
+  },
+  prompt: function (currentData, callback) {
+    console.log('CURR PROFILE: ', currentData);
+    var nameP = {
+      name: 'name',
+      description: 'Name',
+      required: true,
+      default: currentData.name || ''
+    };
+
+    var urlP = {
+      name: 'url',
+      pattern: /[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/,
+      message: 'Must provide a valid url',
+      description: 'URL (Optional)',
+      default: currentData.url || ''
+    }
+
+    var repoP = {
+      name: 'githubUrl',
+      pattern: /[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/,
+      message: 'Must provide a valid url',
+      description: 'GitHub URL (Optional)',
+      default: currentData.githubUrl || ''
+    }
+
+    console.log('currentData.trackOk === true', currentData.trackOk, ' _ ',  currentData.trackOk === true);    
+    var trackP = {
+      name: 'trackOk',
+       description: "Share usage information? (Y/n)",
+       default: currentData.trackOk === true ? 'y' : 'n',
+       pattern: /y|n|yes|no|Y|N/,
+       message: "Please answer y or n."
+    }
+
+    var toPrompt = [
+      nameP, urlP, repoP, trackP
+    ];
+
+    prompt.delimiter = '';
+    prompt.message = 'User info -- ';
+
+    prompt.start();
+    Matrix.loader.stop();
+    prompt.get(toPrompt, function (err, results) {
+
+      if (err) {
+        if (err.toString().indexOf('canceled') > 0) {
+          console.log('');
+          process.exit();
+        } else {
+          console.log("Error: ", err);
+          process.exit();
+        }
+      }
+      
+      console.log('bRESU', results);
+      if (results.trackOk.toLowerCase() === 'y' || results.trackOk.toLowerCase() === 'yes'){
+        results.trackOk = true;
+      } else {
+        results.trackOk = false;
+      }
+      console.log('aRESU', results);
+
+      callback(err, results);
+
+    });
+  },
+  get: function (callback) {
+    Matrix.firebase.user.get(function (err, userProfile) { 
+      callback(err, userProfile);
+    });
+  },
+  update: function (profile, callback) {
+    console.log('UPDATE WITH: ', profile);
+    var updateError;
+    var events = {
+      error: function (err) {
+        Matrix.loader.stop();        
+        updateError = new Error('Error updating user profile: '.red, err.message);
+        callback(updateError);
+      },
+      finished: function () {
+        Matrix.loader.stop();
+        console.log('User profile updated successfully');
+        callback(updateError);
+      },
+      start: function () {
+        Matrix.loader.stop();
+        console.log('User profile update requested...');
+        Matrix.loader.start();
+      },
+      progress: function () {
+        Matrix.loader.stop();
+        console.log('Updating...');
+        Matrix.loader.start();
+      }
+    };
+
+    console.log('PROFILE: ' , profile);    
+    Matrix.firebase.user.update(profile, events);
+  }
+}
+
 module.exports = {
   allowAllPolicy: allowAllPolicy,
   checkAppFolder: checkAppFolder,
@@ -896,6 +1016,7 @@ module.exports = {
   lookupDeviceName: lookupDeviceName,
   packageApp: packageApp,
   patchVersion: patchVersion,
+  profile: profile,
   removeConfig: removeConfig,
   saveConfig: saveConfig,
   updateFile: updateFile,
